requirements.txt
Django>=3.2
psycopg2-binary>=2.8
django-environ>=0.4
django-extensions>=3.1
djangorestframework
django-tgbot
requests